Code refactoring

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate (double weightKG, double growthM) {
-        double index = weightKG / (growthM*growthM);
-        return (int)index;
+    public int calculate(double weightKG, double growthM) {
+        double index = weightKG / (growthM * growthM);
+        return (int) index;
     }
 }
